lm_studio/project_update_5/requirements.txt
faiss-cpu
sentence-transformers
pandas
numpy
pyyaml
requests
tqdm
